requirements.txt
ocrd>=3.0
click>=7.0
scipy
numpy>=1.17.0
pillow>=7.1.2
scikit-image>=0.17.2
